+++ util/stringUtils.ts
@@ -1,0 +1,2 @@
+export const getAcronymFromString = (str: string) =>
+  [...(str.trim().match(/\b(\w)/g) || '')].join('').toUpperCase();
